Add simple delegator for stripe subscription
